style: extract momoized sorting to useSortedProjects
